// src/app/shared/models/sumario.ts
export class Sumario {
    numeroSumario: string;
    anio: string;
    causa: string;
    act_sum:string;
    act_com:string;
    comentarios:string;
    caratula:string;
    status: number;
    created_by: string;
    updated_by: string;
    fecha_inicio: Date;
    fecha_fin: Date;
    juzgado_id: number;
    tecnico_id: number;
    estado_id: number;
}